requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
pandas>=1.1.3
matplotlib>=3.3.2
scikit-learn>=0.23.2
tqdm>=4.50.2
Pillow>=8.0.0 